arch-review: migrate near-identical file-comment functions into useFileComments
Second half of finding 5: move the near-identical functions into the
composable, reconciling each difference deliberately (not blindly picking
one side):

- computeSelectionLines: keep FileViewerModal's explanatory comments;
  source content comes from the injected content() getter.
- updateSelectionAnchorFromRange: adopt PinnedFilePanel's left clamp
  (keeps the trigger pill inside a narrow panel; a no-op in the wide
  modal), parameterized by containerEl.
- onSelectionChange: keep FileViewerModal's comments; listeners in both
  components now point at comments.onSelectionChange.
- openCommentForSelection / openCommentForCsvCell: drop the dead 'anchor'
  field on FileViewerModal's draft; close the hover-pin read popover via
  the injected closeReadPopover callback.
- saveComment: keep FileViewerModal's lastSelectionRange reset (the
  selection is gone after save; consistent with cancelComment).
- saveEditComment: keep FileViewerModal's highlight reapply for symmetry
  with the delete/save paths.
- deleteFileComment: unify on PinnedFilePanel's single-arg form (path from
  the injected path() getter; equals the comment's stored path).
- handleEditImageUpload / onCsvCellActivate: only the path / comments
  source differed; both now use the injected getters.

Genuinely divergent functions stay in the components (do not force them):
applyHighlights (different guard sets), scrollToHighlight (single popup vs
multi-fragment hover-pin), onScrollReanchor (PinnedFilePanel suppresses
during programmatic scrolls and closes its read popover). They reference
the composable's state/helpers through comments.*.

FileViewerModal: 2938 -> 2181 lines (-757). PinnedFilePanel: 2241 -> 1748
(-493). useFileComments.ts: 665 lines. Tests (36 files / 298 tests),
vue-tsc --noEmit, and eslint all pass.
