+++ web/src/components/FileViewerModal.vue
@@ -405,11 +405,10 @@ import { openWorkspaceFileExternally } from '../lib/openWorkspaceFile'
 import { createTerminalDiffLines, terminalDiffPrefix, type TerminalDiffKind } from '../lib/terminalDiff'
 import { isCsvPath } from '../lib/csv'
 import { askConfirm } from '../lib/confirm'
-import { useFileComments, type LineRange } from '../composables/useFileComments'
+import { useFileComments } from '../composables/useFileComments'
 import CommentComposePopover from './CommentComposePopover.vue'
 const ExcalidrawViewer = defineAsyncComponent(() => import('./ExcalidrawViewer.vue'))
 const CsvViewer = defineAsyncComponent(() => import('./CsvViewer.vue'))
-import type { CsvCellRef } from './CsvViewer.vue'
 
 const store = useFileViewerStore()
 const projectsStore = useProjectStore()
@@ -644,12 +643,6 @@ function isUrl(value: string): boolean {
 const activeFileComments = computed(() =>
   projectsStore.fileCommentsFor(cleanPath(store.path))
 )
-
-
-function deleteFileComment(path: string, id: string): void {
-  projectsStore.removeFileComment(path, id)
-  nextTick(() => applyHighlights())
-}
 
 // ── Pre line rendering & line-based highlighting ───────────────────
 const contentLines = computed(() => {
@@ -921,7 +914,7 @@ function deletePopupComment(): void {
   if (!id) return
   activePopupId.value = null
   const c = activeFileComments.value.find(x => x.id === id)
-  if (c) deleteFileComment(c.path, id)
+  if (c) comments.deleteFileComment(id)
 }
 
 function editFromPopup(c: { id: string; comment: string; images?: string[] }): void {
@@ -1067,6 +1060,7 @@ const comments = useFileComments({
   mdEl,
   preEl,
   preCodeEl,
+  scrollToHighlight,
 })
 // Script code reaches the shared methods/state through `comments.*`; only
 // the template-bound names are destructured here.
@@ -1074,8 +1068,9 @@ const {
   selectionAnchor, draftAnchor, commentDraft, composeText, csvCellComments,
   commentDraftImages, editingCommentId, editDraftText, editingCommentImages, editAnchor,
   isHighlightedLine, commentIdForLine, commentLineLabel,
-  onCsvCellSelect, cancelComment, handleDraftImageUpload, removeDraftImage,
-  cancelEditComment, removeEditImage,
+  onCsvCellSelect, onCsvCellActivate, openCommentForSelection, openCommentForCsvCell,
+  cancelComment, saveComment, handleDraftImageUpload, removeDraftImage,
+  cancelEditComment, saveEditComment, handleEditImageUpload, removeEditImage,
 } = comments
 comments.setApplyHighlights(applyHighlights)
 
@@ -1173,88 +1168,9 @@ function cleanPath(p: string): string {
   return p.replace(/:\d+$/, '')
 }
 
-// Compute a {start, end} source line range for the active selection.
-// Two strategies:
-//   • <pre> branch: text-node offsets map 1:1 onto store.content, so we
-//     read range.startOffset/endOffset directly via charOffsetFrom and
-//     count newlines. This is exact.
-//   • markdown branch: rendered DOM doesn't map onto source, so we
-//     substring-search the source markdown for the first ~50 chars (start
-//     line) and the last ~50 chars (end line) of the selection. Falls back
-//     to a single line when the second search misses.
-function computeSelectionLines(range: Range, selectionText: string): LineRange {
-  const src = store.content
-  if (!src) return null
-
-  // Plain-text branch: exact mapping via offsets into the <code> root.
-  const codeRoot = preCodeEl.value
-  if (codeRoot && codeRoot.contains(range.startContainer)) {
-    const startOff = comments.charOffsetFrom(codeRoot, range.startContainer, range.startOffset)
-    const endOff = comments.charOffsetFrom(codeRoot, range.endContainer, range.endOffset)
-    if (startOff != null && endOff != null) {
-      const a = Math.min(startOff, endOff)
-      const b = Math.max(startOff, endOff)
-      const start = comments.lineAt(src, a)
-      // For end, look at the last char of the selection (b - 1) so a
-      // selection ending at the start of a line doesn't bleed into it.
-      const end = b > a ? comments.lineAt(src, b - 1) : start
-      return { start, end: Math.max(end, start) }
-    }
-  }
-
-  // Markdown branch: best-effort substring lookup.
-  const trimmed = selectionText.trim()
-  if (!trimmed) return null
-  const head = trimmed.slice(0, 60)
-  let startIdx = src.indexOf(head)
-  if (startIdx === -1) {
-    // Try a shorter prefix from the first line of the rendered selection,
-    // which is usually the most stable token to find in source.
-    const firstLine = trimmed.split(/\n/, 1)[0].trim().slice(0, 30)
-    if (firstLine.length >= 4) startIdx = src.indexOf(firstLine)
-  }
-  if (startIdx === -1) return null
-  const start = comments.lineAt(src, startIdx)
-
-  const tail = trimmed.slice(-60).trim()
-  if (tail.length >= 4 && tail !== head) {
-    // Search starting after the head match so identical phrases earlier in
-    // the doc don't pull the end line backwards.
-    const tailIdx = src.indexOf(tail, startIdx)
-    if (tailIdx !== -1) {
-      const end = comments.lineAt(src, tailIdx + tail.length - 1)
-      return { start, end: Math.max(end, start) }
-    }
-  }
-  return { start, end: start }
-}
-
-function updateSelectionAnchorFromRange(range: Range): void {
-  const modal = modalEl.value
-  const body = bodyEl.value
-  if (!modal || !body) {
-    comments.selectionAnchor.value = null
-    return
-  }
-
-  const rects = range.getClientRects()
-  const endRect = rects.length ? rects[rects.length - 1] : range.getBoundingClientRect()
-  const bodyRect = body.getBoundingClientRect()
-  const visible = endRect.bottom > bodyRect.top
-    && endRect.top < bodyRect.bottom
-    && endRect.right > bodyRect.left
-    && endRect.left < bodyRect.right
-  if (!visible) {
-    comments.selectionAnchor.value = null
-    return
-  }
-
-  const modalRect = modal.getBoundingClientRect()
-  const top = endRect.bottom - modalRect.top + 2
-  const left = Math.max(8, endRect.right - modalRect.left + 6)
-  comments.selectionAnchor.value = { top, left }
-}
-
+// Re-anchor the floating comment trigger on scroll. Kept here (not in
+// useFileComments): PinnedFilePanel's variant additionally suppresses
+// re-anchoring during programmatic scrolls and closes its hover-pin popover.
 function onScrollReanchor(): void {
   if (comments.commentDraft.value || !comments.lastSelectionRange) return
   try {
@@ -1263,166 +1179,15 @@ function onScrollReanchor(): void {
       comments.selectionAnchor.value = null
       return
     }
-    updateSelectionAnchorFromRange(comments.lastSelectionRange)
+    comments.updateSelectionAnchorFromRange(comments.lastSelectionRange)
   } catch {
     comments.lastSelectionRange = null
     comments.selectionAnchor.value = null
   }
 }
 
-function onSelectionChange(): void {
-  if (!isCommentable.value) {
-    comments.lastSelectionRange = null
-    comments.selectionAnchor.value = null
-    return
-  }
-  // While the comment composer is open the selection has been "captured" —
-  // don't keep retracking it (the textarea steals focus and would clear it).
-  if (comments.commentDraft.value) return
-  const sel = window.getSelection()
-  if (!sel || sel.isCollapsed || sel.rangeCount === 0) {
-    comments.lastSelectionRange = null
-    comments.selectionAnchor.value = null
-    return
-  }
-  const range = sel.getRangeAt(0)
-  // Only react to selections inside the rendered file view — selecting text
-  // in the path subtitle or modal chrome shouldn't trigger the comment UI.
-  // Either the markdown div or the <pre> code branch counts.
-  const targets: (HTMLElement | undefined)[] = [mdEl.value, preEl.value, preCodeEl.value]
-  const inside = targets.some(
-    el => el && el.contains(range.startContainer) && el.contains(range.endContainer)
-  )
-  if (!inside) {
-    comments.lastSelectionRange = null
-    comments.selectionAnchor.value = null
-    return
-  }
-  const text = sel.toString().trim()
-  if (!text) {
-    comments.lastSelectionRange = null
-    comments.selectionAnchor.value = null
-    return
-  }
-  comments.lastSelectionText = text
-  comments.lastSelectionLines = computeSelectionLines(range, text)
-  comments.lastSelectionRange = range.cloneRange()
-  updateSelectionAnchorFromRange(range)
-}
-
-function openCommentForSelection(): void {
-  if (!comments.selectionAnchor.value || !comments.lastSelectionText) return
-  comments.draftAnchor.value = comments.toViewportAnchor(comments.selectionAnchor.value)
-  comments.commentDraft.value = {
-    selection: comments.lastSelectionText,
-    text: '',
-    lines: comments.lastSelectionLines,
-    cell: null,
-  }
-  comments.commentDraftImages.value = []
-  comments.selectionAnchor.value = null
-  comments.lastSelectionRange = null
-  comments.lastCsvCell = null
-  window.getSelection()?.removeAllRanges()
-  nextTick(() => applyHighlights())
-}
-
-function onCsvCellActivate(cell: CsvCellRef): void {
-  const match = activeFileComments.value.find(
-    c => c.lineStart === cell.row && c.colIndex === cell.colIndex,
-  )
-  if (match) scrollToHighlight(match.id)
-}
-
-function openCommentForCsvCell(): void {
-  if (!comments.selectionAnchor.value || !comments.lastCsvCell) return
-  comments.draftAnchor.value = comments.toViewportAnchor(comments.selectionAnchor.value)
-  comments.commentDraft.value = {
-    selection: comments.lastCsvCell.value,
-    text: '',
-    lines: { start: comments.lastCsvCell.row, end: comments.lastCsvCell.row },
-    cell: {
-      row: comments.lastCsvCell.row,
-      colIndex: comments.lastCsvCell.colIndex,
-      colHeader: comments.lastCsvCell.colHeader,
-    },
-  }
-  comments.commentDraftImages.value = []
-  comments.selectionAnchor.value = null
-  comments.lastSelectionRange = null
-  nextTick(() => applyHighlights())
-}
-
-function saveComment(): void {
-  const draft = comments.commentDraft.value
-  if (!draft) return
-  const note = draft.text.trim()
-  if (!note) return
-  projectsStore.addPendingComment({
-    path: cleanPath(store.path),
-    selection: draft.selection,
-    comment: note,
-    lineStart: draft.lines?.start ?? null,
-    lineEnd: draft.lines?.end ?? null,
-    colIndex: draft.cell?.colIndex ?? null,
-    colHeader: draft.cell?.colHeader ?? null,
-    images: comments.commentDraftImages.value.length ? comments.commentDraftImages.value : undefined,
-  })
-  comments.commentDraft.value = null
-  comments.draftAnchor.value = null
-  comments.commentDraftImages.value = []
-  comments.lastSelectionText = ''
-  comments.lastSelectionLines = null
-  comments.lastSelectionRange = null
-  comments.lastCsvCell = null
-  nextTick(() => applyHighlights())
-}
-
-// ── Edit existing comment ────────────────────────────────────────────
-function saveEditComment(id: string | null): void {
-  if (!id) return
-  const note = comments.editDraftText.value.trim()
-  if (!note) return
-  const path = cleanPath(store.path)
-  projectsStore.updateFileComment(path, id, note)
-  // Sync images: remove existing ones that are gone, add new ones
-  const existing = projectsStore.fileCommentsFor(path).find(c => c.id === id)
-  const existingImages = existing?.images || []
-  const nextImages = comments.editingCommentImages.value
-  for (const img of existingImages) {
-    if (!nextImages.includes(img)) projectsStore.removeFileCommentImage(path, id, img)
-  }
-  for (const img of nextImages) {
-    if (!existingImages.includes(img)) projectsStore.addFileCommentImage(path, id, img)
-  }
-  comments.cancelEditComment()
-  nextTick(() => applyHighlights())
-}
-
-async function handleEditImageUpload(e: Event, id: string | null): Promise<void> {
-  if (!id) return
-  const input = e.target as HTMLInputElement
-  if (!input.files?.length) return
-  const chatId = projectsStore.activeChatId
-  if (!chatId) return
-  try {
-    const refs = await projectsStore.uploadImageRefs(chatId, Array.from(input.files))
-    const path = cleanPath(store.path)
-    for (const ref of refs) {
-      projectsStore.addFileCommentImage(path, id, ref)
-    }
-    // Refresh local edit state from store
-    const c = projectsStore.fileCommentsFor(path).find(x => x.id === id)
-    if (c?.images) comments.editingCommentImages.value = [...c.images]
-  } catch (err) {
-    console.error('Comment image upload failed:', err)
-  }
-  input.value = ''
-}
-
-
 if (typeof document !== 'undefined') {
-  document.addEventListener('selectionchange', onSelectionChange)
+  document.addEventListener('selectionchange', comments.onSelectionChange)
 }
 function onSidebarResize(): void {
   updateCommentPositions()
@@ -1440,7 +1205,7 @@ if (typeof window !== 'undefined') {
 }
 onBeforeUnmount(() => {
   if (typeof document !== 'undefined') {
-    document.removeEventListener('selectionchange', onSelectionChange)
+    document.removeEventListener('selectionchange', comments.onSelectionChange)
   }
   if (typeof window !== 'undefined') {
     window.removeEventListener('resize', onSidebarResize)
